Make agents have their own type of pheromone.
This should induce some interesting behaviours of
the simulation as agents now form groups, where each
agent obeys to their group's type of the pheromone.

Also, due to the algorithm of pheromone detection,
if pheromones of different agents are similar,
agents will drift towards them slightly.

The pheromone type is defined as an RGB color.

The Pheromone level is detected by the mean value of
pheromone levels in the 3x3 square around the
sensor, which is then multiplied by the pheromone
type of the agent (RGB, 3D vector).